requirements.txt
colorama
opencv-python-headless
db-dtypes
httpx
ipdb
ipykernel
mediapipe
nbresult
numpy==1.23.5
pandas==1.5.3
pillow
protobuf
psycopg2-binary
pyarrow
pytest
pytest-asyncio
python-dotenv
scikit-learn==1.2.1
wheel
xgboost==1.6.2
yapf

fastapi
pytz
uvicorn
python-multipart
